Softly disable dm-verity
We are going to be shipping system images to be flashed via recovery
as full OTAs instead of AOSP (or Google factory) images and most
people will end up flashing GApps, SU and whatnot on top. The verity
information is not going to be relevant. If we decide to ship with a
closed off system, this commit should be reverted.

Change-Id: I1d9c7ea9fa4e7a4b8acd10a85005db1591b51b1a