Add /vendor/ proprietary blobs
We will include the vendor partition along with the system partition
to make sure the user doesn't miss out on any blobs. Old-school builds
that had the same partition for /system/vendor stuff already did this
and we can do too, to make sure nothing goes wrong with them blobs.

Change-Id: I5632991aafb29644465d1e738abb368b3a9a1c36